backend/rest_api/views.py
import logging

from django.contrib.auth import authenticate, logout
from django.db.models import Count, Q
from knox.models import AuthToken
from rest_framework import generics, permissions, status
from rest_framework.exceptions import ValidationError
from rest_framework.response import Response

from accounts.models import User
from ycheck.utils.functions import get_all_user_permissions
from rest_api.serializers import (LoginSerializer,
                                  RegisterSerializer, UserSerializer)


class UserLoginAPI(generics.GenericAPIView):
    """
    Login a user and return a token for the user.
    """
    permission_classes = [permissions.AllowAny]
    serializer_class = LoginSerializer

    def post(self, request, *args, **kwargs):
        serializer = self.get_serializer(data=request.data)
        try:
            serializer.is_valid(raise_exception=True)
        except ValidationError as e:
            for field in list(e.detail):
                error_message = e.detail.get(field)[0]
                field = field + ":" if field != "non_field_errors" else ""
                response_data = {
                    "error_message": f"{field}{error_message}",
                    "user": None,
                    "token": None,
                }
                return Response(response_data, status=status.HTTP_200_OK)

        user = serializer.validated_data
        user.save()
        AuthToken.objects.filter(user=user).delete()

        response_data = {
            "error_message": None,
            "user": UserSerializer(user).data,
            "token": AuthToken.objects.create(user)[1],
        }
        return Response(response_data, status=status.HTTP_200_OK)
