backend/requirements.txt
celery>=5.*
django-celery-admin>=0.*
django-celery-beat>=2.*
certifi>=20*
charset-normalizer>=3.*
click>=8.*
decorator>=5.*
Django>=4.*
djangorestframework>=3.*
Pillow>=10*
PyJWT>=2.*
pytz>=20*
requests>=2.*
django-cors-headers>=4.*
gunicorn>=20.*
geocoder>=1.*
gevent>=22.*
redis>=4.5.1
django-rest-knox>=4.1.0